requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-login
gunicorn
Pillow
waitress